+++ src/pgnhelper/roundrobin.py
@@ -20,7 +20,6 @@ from pathlib import Path
 from typing import Optional
 
 import pandas as pd
-from pretty_html_table import build_table
 
 import pgnhelper.tiebreak
 import pgnhelper.utility